Extract ObjectDetailsTable.styled
